Change data type of "rate" to double AND add OBJECT_LEN constant

+++ NORM/NormDefinition.h
@@ -26,7 +26,7 @@ namespace DCCast {
 
     union DCCommandContent {
         struct {
-            uint32_t rate;
+            double rate;
         } updateRate;
     };
